# Conflicts:
#	pom.xml
#	redisson-all/pom.xml
#	redisson-tomcat/pom.xml
#	redisson-tomcat/redisson-tomcat-6/pom.xml
#	redisson-tomcat/redisson-tomcat-7/pom.xml
#	redisson-tomcat/redisson-tomcat-8/pom.xml
#	redisson/pom.xml

+++ redisson/src/main/java/org/redisson/reactive/RedissonLexSortedSetReactive.java
@@ -25,6 +25,11 @@ import org.redisson.client.codec.StringCodec;
 import org.redisson.client.protocol.RedisCommands;
 import org.redisson.command.CommandReactiveExecutor;
 
+/**
+ * 
+ * @author Nikita Koksharov
+ *
+ */
 public class RedissonLexSortedSetReactive extends RedissonScoredSortedSetReactive<String> implements RLexSortedSetReactive {
 
     public RedissonLexSortedSetReactive(CommandReactiveExecutor commandExecutor, String name) {
@@ -32,7 +37,7 @@ public class RedissonLexSortedSetReactive extends RedissonScoredSortedSetReactiv
     }
 
     @Override
-    public Publisher<Long> addAll(Publisher<? extends String> c) {
+    public Publisher<Integer> addAll(Publisher<? extends String> c) {
         return new PublisherAdder<String>(this).addAll(c);
     }
 
@@ -130,12 +135,12 @@ public class RedissonLexSortedSetReactive extends RedissonScoredSortedSetReactiv
     }
 
     @Override
-    public Publisher<Long> add(String e) {
+    public Publisher<Integer> add(String e) {
         return commandExecutor.writeReactive(getName(), StringCodec.INSTANCE, RedisCommands.ZADD_RAW, getName(), 0, e);
     }
 
     @Override
-    public Publisher<Long> addAll(Collection<? extends String> c) {
+    public Publisher<Integer> addAll(Collection<? extends String> c) {
         List<Object> params = new ArrayList<Object>(2*c.size());
         for (Object param : c) {
             params.add(0);
